requirements.txt
streamlit
tensorflow
numpy
pillow
matplotlib
seaborn
scikit-learn
streamlit-option-menu